refactor(broker): extract Order domain types into order_types.rs
Moves OrderStatus, OrderStateChange, Order (struct + impls + Display),
CompressionType, and the order_from_request / skipped_order_from_request
/ proving_order_from_request helpers out of the 1900-line lib.rs into
their own module.

Order's fields become pub(crate) (they were module-private when Order
lived in lib.rs and got that access for free across the crate; the
explicit pub(crate) preserves that scope from the new module location).

format_order_id stays private to order_types. format_expiries stays in
lib.rs but becomes pub(crate) so Order's Display impl can reach it.

The journal_bytes_backwards_compat unit test moves with the Order type
into order_types::tests where it now lives.

lib.rs imports trimmed (Bytes, FixedBytes, OnceLock, the selector
helpers, and serde::ts_seconds are no longer needed at the top level).
Public types are re-exported via `pub use`; the pub(crate) helpers are
re-exported via `pub(crate) use` so existing `crate::Order`,
`crate::proving_order_from_request`, etc. call-sites continue to work.

Net: lib.rs shrinks by ~225 lines. No semantic change.
